+++ crates/astria-sequencer-types/src/lib.rs
@@ -1,9 +1,11 @@
+pub mod abci_codes;
 pub(crate) mod namespace;
 pub mod sequencer_block_data;
 pub mod serde;
 pub mod tendermint;
 pub mod test_utils;
 
+pub use abci_codes::AbciCode;
 pub use namespace::{
     Namespace,
     DEFAULT_NAMESPACE,
