+++ crates/astria-sequencer/src/service/info/mod.rs
@@ -33,7 +33,8 @@ use tracing::{
 
 mod abci_query_router;
 
-use super::AbciCode;
+use sequencer_types::abci_codes::AbciCode;
+
 use crate::state_ext::StateReadExt;
 
 #[derive(Clone)]
